+++ starlette/schemas.py
@@ -8,7 +8,7 @@ from starlette.routing import BaseRoute, Mount, Route
 
 try:
     import yaml
-except ImportError:  # pragma: nocover
+except ModuleNotFoundError:  # pragma: nocover
     yaml = None  # type: ignore[assignment]
 
 
@@ -132,7 +132,6 @@ class SchemaGenerator(BaseSchemaGenerator):
         endpoints_info = self.get_endpoints(routes)
 
         for endpoint in endpoints_info:
-
             parsed = self.parse_docstring(endpoint.func)
 
             if not parsed: