+++ starlette/templating.py
@@ -18,7 +18,7 @@ try:
         pass_context = jinja2.pass_context
     else:  # pragma: nocover
         pass_context = jinja2.contextfunction  # type: ignore[attr-defined]
-except ImportError:  # pragma: nocover
+except ModuleNotFoundError:  # pragma: nocover
     jinja2 = None  # type: ignore[assignment]
 
 
@@ -64,23 +64,31 @@ class Jinja2Templates:
 
     def __init__(
         self,
-        directory: typing.Union[str, PathLike],
+        directory: typing.Union[
+            str, PathLike, typing.Sequence[typing.Union[str, PathLike]]
+        ],
         context_processors: typing.Optional[
             typing.List[typing.Callable[[Request], typing.Dict[str, typing.Any]]]
         ] = None,
-        **env_options: typing.Any
+        **env_options: typing.Any,
     ) -> None:
         assert jinja2 is not None, "jinja2 must be installed to use Jinja2Templates"
         self.env = self._create_env(directory, **env_options)
         self.context_processors = context_processors or []
 
     def _create_env(
-        self, directory: typing.Union[str, PathLike], **env_options: typing.Any
+        self,
+        directory: typing.Union[
+            str, PathLike, typing.Sequence[typing.Union[str, PathLike]]
+        ],
+        **env_options: typing.Any,
     ) -> "jinja2.Environment":
         @pass_context
-        def url_for(context: dict, name: str, **path_params: typing.Any) -> URL:
+        # TODO: Make `__name` a positional-only argument when we drop Python 3.7
+        # support.
+        def url_for(context: dict, __name: str, **path_params: typing.Any) -> URL:
             request = context["request"]
-            return request.url_for(name, **path_params)
+            return request.url_for(__name, **path_params)
 
         loader = jinja2.FileSystemLoader(directory)
         env_options.setdefault("loader", loader)