+++ starlette/types.py
@@ -1,5 +1,7 @@
 import typing
 
+AppType = typing.TypeVar("AppType")
+
 Scope = typing.MutableMapping[str, typing.Any]
 Message = typing.MutableMapping[str, typing.Any]
 
@@ -8,8 +10,8 @@ Send = typing.Callable[[Message], typing.Awaitable[None]]
 
 ASGIApp = typing.Callable[[Scope, Receive, Send], typing.Awaitable[None]]
 
-StatelessLifespan = typing.Callable[[object], typing.AsyncContextManager[typing.Any]]
-StateLifespan = typing.Callable[
-    [typing.Any, typing.Dict[str, typing.Any]], typing.AsyncContextManager[typing.Any]
+StatelessLifespan = typing.Callable[[AppType], typing.AsyncContextManager[None]]
+StatefulLifespan = typing.Callable[
+    [AppType], typing.AsyncContextManager[typing.Mapping[str, typing.Any]]
 ]
-Lifespan = typing.Union[StatelessLifespan, StateLifespan]
+Lifespan = typing.Union[StatelessLifespan[AppType], StatefulLifespan[AppType]]