+++ tests/test_testclient.py
@@ -1,4 +1,5 @@
 import itertools
+import sys
 from asyncio import current_task as asyncio_current_task
 from contextlib import asynccontextmanager
 
@@ -43,9 +44,6 @@ def current_task():
 
 def startup():
     raise RuntimeError()
-
-
-startup_error_app = Starlette(on_startup=[startup])
 
 
 def test_use_testclient_in_endpoint(test_client_factory):
@@ -166,6 +164,11 @@ def test_use_testclient_as_contextmanager(test_client_factory, anyio_backend_nam
 
 
 def test_error_on_startup(test_client_factory):
+    with pytest.deprecated_call(
+        match="The on_startup and on_shutdown parameters are deprecated"
+    ):
+        startup_error_app = Starlette(on_startup=[startup])
+
     with pytest.raises(RuntimeError):
         with test_client_factory(startup_error_app):
             pass  # pragma: no cover
@@ -271,3 +274,48 @@ def test_query_params(test_client_factory, param: str):
     client = test_client_factory(app)
     response = client.get("/", params={"param": param})
     assert response.text == param
+
+
+@pytest.mark.parametrize(
+    "domain, ok",
+    [
+        pytest.param(
+            "testserver",
+            True,
+            marks=[
+                pytest.mark.xfail(
+                    sys.version_info < (3, 11),
+                    reason="Fails due to domain handling in http.cookiejar module (see "
+                    "#2152)",
+                ),
+            ],
+        ),
+        ("testserver.local", True),
+        ("localhost", False),
+        ("example.com", False),
+    ],
+)
+def test_domain_restricted_cookies(test_client_factory, domain, ok):
+    """
+    Test that test client discards domain restricted cookies which do not match the
+    base_url of the testclient (`http://testserver` by default).
+
+    The domain `testserver.local` works because the Python http.cookiejar module derives
+    the "effective domain" by appending `.local` to non-dotted request domains
+    in accordance with RFC 2965.
+    """
+
+    async def app(scope, receive, send):
+        response = Response("Hello, world!", media_type="text/plain")
+        response.set_cookie(
+            "mycookie",
+            "myvalue",
+            path="/",
+            domain=domain,
+        )
+        await response(scope, receive, send)
+
+    client = test_client_factory(app)
+    response = client.get("/")
+    cookie_set = len(response.cookies) == 1
+    assert cookie_set == ok